@@ -12,9 +12,9 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=8.0"
     },
     "autoload": {
-        "psr-0": { "peterurk\\Cipher": "src" }
+        "psr-4": { "peterurk\\Cipher\\": "src/" }
     }
 }
